# openapi/synapse.yaml
openapi: 3.1.0
info:
  title: Synapse API
  version: 0.1.0
  summary: Organizational Brain service for AI agents and team workflows.
  description: |
    Synapse is the governed layer that gives agents and teams permission-aware
    retrieval, a versioned skill registry, bounded memory, a user/team/org
    context service, a policy-guarded tool/connector (MCP) gateway, and durable
    workflow orchestration.

    Production identity is carried by a verified bearer JWT (JWKS, static RS256,
    or HS256). Trusted X-Principal-Id/X-Tenant-Id headers remain available for
    development or a trusted identity-aware gateway. Every governed endpoint is
    attributable for audit.

  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: ResQ Software
    url: https://github.com/resq-software/synapse
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /tool.execute:
    post:
      tags: [tools]
      operationId: toolExecute
      summary: Execute a tool/connector call.
      description: |
        Executes an enabled tenant-owned tool contract through the outbound MCP gateway.
        Synapse validates arguments and connector scopes before dispatch. Registry or
        request policy may require durable approval before any autonomous write.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolExecuteRequest'
      responses:
        '200':
          description: Tool executed (or queued for approval).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolExecuteResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tools.register:
    post:
      tags: [tools]
      operationId: toolsRegister
      summary: Create or update a tenant tool contract.
      description: Admin-only. Validates the input schema and increments the contract revision.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolRegisterRequest'
      responses:
        '200':
          description: Registered contract.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolRegisterResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tools.list:
    post:
      tags: [tools]
      operationId: toolsList
      summary: List tenant tool contracts.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      responses:
        '200':
          description: Contracts ordered by tool id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tools.decide:
    post:
      tags: [tools]
      operationId: toolsDecide
      summary: Approve or deny a standalone pending execution.
      description: Admin-only. Approval revalidates the current contract before connector dispatch.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolDecisionRequest'
      responses:
        '200':
          description: Decision applied or terminal outcome replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolLifecycleResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tools.rollback:
    post:
      tags: [tools]
      operationId: toolsRollback
      summary: Execute registered compensation for an execution.
      description: Admin-only and idempotent; at most one compensation execution per original.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolRollbackRequest'
      responses:
        '200':
          description: Compensation outcome.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolLifecycleResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified JWT; required when Synapse is configured for JWT authentication.
  parameters:
    PrincipalId:
      name: X-Principal-Id
      in: header
      required: false
      description: Trusted-header identity for development/gateway mode; ignored when JWT auth is configured.
      schema:
        type: string
      example: alice@resq
    TenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: Owning tenant/organization for this request.
      schema:
        type: string
      example: resq
    TeamIds:
      name: X-Team-Ids
      in: header
      required: false
      description: Comma-separated team ids the principal is acting within.
      schema:
        type: string
      example: team-ops,team-air
    Role:
      name: X-Role
      in: header
      required: false
      description: Platform/business role of the principal, used for RBAC.
      schema:
        type: string
      example: operator
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid bearer token or trusted principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted by policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The operation conflicts with the resource's current lifecycle state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: Stable error envelope returned for all failure responses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              examples: [not_found, bad_request, unauthorized, forbidden, db_error, internal]
            message:
              type: string
              description: Human-readable error message.
...
    ToolPolicy:
      type: object
      description: Policy block on a tool call.
      properties:
        approval_mode:
          type: string
          description: Whether the tool call requires human approval before running.
          enum: [none, required]
          default: none
        reason:
          type: string
          description: Human-readable justification for the call/approval.
    ToolExecuteRequest:
      type: object
      description: A policy-guarded tool/connector execution request.
      required: [tenant_id, principal_id, tool_id]
      properties:
        tenant_id:
          type: string
          example: resq
        principal_id:
          type: string
          example: alice@resq
        tool_id:
          type: string
          description: Identifier of the tool/connector to execute.
          example: jira.create_issue
        arguments:
          type: object
          description: Arguments passed to the tool (validated against its input schema).
          additionalProperties: true
        policy:
          $ref: '#/components/schemas/ToolPolicy'
    ToolExecuteResponse:
      type: object
      description: Result of a tool execution (or approval gating).
      required: [tool_id, status]
      properties:
        tool_id:
          type: string
        execution_id:
          type: string
          format: uuid
          description: Durable execution intent id, including pending approvals.
        status:
          type: string
          description: Execution status (for example, executed or pending).
          example: pending
        output:
          type: object
          description: Tool output, when executed.
          additionalProperties: true
        requires_approval:
          type: boolean
          description: True if the call is gated on human approval.
          default: false
    ToolDefinition:
      type: object
      required: [tool_id, input_schema, required_scopes, approval_mode, enabled, revision]
      properties:
        tool_id:
          type: string
          maxLength: 255
        description:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        required_scopes:
          type: array
          uniqueItems: true
          items:
            type: string
        approval_mode:
          type: string
          enum: [none, required]
        rollback_tool_id:
          type: [string, 'null']
        enabled:
          type: boolean
        revision:
          type: integer
          format: int64
          minimum: 1
    ToolRegisterRequest:
      type: object
      required: [tool_id]
      properties:
        tool_id:
          type: string
          maxLength: 255
        description:
          type: string
          default: ''
        input_schema:
          type: object
          default: { type: object }
          additionalProperties: true
        required_scopes:
          type: array
          uniqueItems: true
          items:
            type: string
        approval_mode:
          type: string
          enum: [none, required]
          default: required
        rollback_tool_id:
          type: [string, 'null']
        enabled:
          type: boolean
          default: false
    ToolRegisterResponse:
      type: object
      required: [tool, status]
      properties:
        tool:
          $ref: '#/components/schemas/ToolDefinition'
        status:
          type: string
          const: registered
    ToolListResponse:
      type: object
      required: [tools]
      properties:
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolDefinition'
    ToolDecisionRequest:
      type: object
      required: [execution_id, decision]
      properties:
        execution_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approve, deny]
        reason:
          type: [string, 'null']
    ToolRollbackRequest:
      type: object
      required: [execution_id]
      properties:
        execution_id:
          type: string
          format: uuid
        reason:
          type: [string, 'null']
    ToolLifecycleResponse:
      type: object
      required: [execution_id, tool_id, status, output]
      properties:
        execution_id:
          type: string
          format: uuid
        tool_id:
          type: string
        status:
          type: string
          enum: [pending, approved, denied, executed, failed]
        output:
          description: Stored tool result or compensation result.
...